FIX: database url fix
