+++ src/main/java/com/switching/study_matching_site/domain/Notice.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 public class Notice {
 
     @Id
-    @GeneratedValue
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "NOTICE_ID")
     private Long id;
