--- src/main/java/com/switching/study_matching_site/domain/OfflineStatus.java
@@ -1,5 +1,0 @@
-package com.switching.study_matching_site.domain;
-
-public enum OfflineStatus {
-    OFFLINE, ONLINE
-}
