+++ src/main/java/com/switching/study_matching_site/domain/Participation.java
@@ -1,5 +1,7 @@
 package com.switching.study_matching_site.domain;
 
+import com.switching.study_matching_site.domain.type.EnterStatus;
+import com.switching.study_matching_site.domain.type.RoleType;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
