+++ src/main/java/com/switching/study_matching_site/domain/type/EnterStatus.java
@@ -1,0 +1,5 @@
+package com.switching.study_matching_site.domain.type;
+
+public enum EnterStatus {
+    ENTER, OUT
+}
