+++ src/main/java/com/switching/study_matching_site/domain/type/RequestStatus.java
@@ -1,4 +1,4 @@
-package com.switching.study_matching_site.domain;
+package com.switching.study_matching_site.domain.type;
 
 public enum RequestStatus {
     PENDING,   // 대기 상태
