+++ src/main/java/com/switching/study_matching_site/domain/type/RoleType.java
@@ -1,0 +1,5 @@
+package com.switching.study_matching_site.domain.type;
+
+public enum RoleType {
+    ADMIN, USER
+}
