+++ src/main/java/com/switching/study_matching_site/domain/type/RoomStatus.java
@@ -1,0 +1,5 @@
+package com.switching.study_matching_site.domain.type;
+
+public enum RoomStatus {
+    ON, OFF
+}
