+++ src/main/java/com/switching/study_matching_site/domain/type/TechSkill.java
@@ -1,4 +1,4 @@
-package com.switching.study_matching_site.domain;
+package com.switching.study_matching_site.domain.type;
 
 public enum TechSkill {
     JAVA, PYTHON, KOTLIN, C, JAVASCRIPT
