+++ src/main/java/com/switching/study_matching_site/dto/profile/ProfileUpdateDto.java
@@ -2,7 +2,10 @@ package com.switching.study_matching_site.dto.profile;
 
 import com.switching.study_matching_site.annotation.ValidRegion;
 import com.switching.study_matching_site.domain.*;
-import io.swagger.v3.oas.annotations.Hidden;
+import com.switching.study_matching_site.domain.type.Goal;
+import com.switching.study_matching_site.domain.type.OfflineStatus;
+import com.switching.study_matching_site.domain.type.Region;
+import com.switching.study_matching_site.domain.type.TechSkill;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
